refactor : apiUtils 파일을 새 위치로 이동하고 CartItem 및 Product 관련 타입 정의 추가
