+++ src/component/ProductContainer/ProductContainer.tsx
@@ -1,4 +1,5 @@
-import { CartItemType, ProductType } from "../../constants";
+import { CartItemType } from "../../types/cartItem";
+import { ProductType } from "../../types/product";
 import Product from "../Product/Product";
 import { ProductContainerLayout } from "./ProductContainer.style";
